requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pillow>=9.0.0
jupyter>=1.0.0
streamlit>=1.28.0